+++ vgo-core/src/test/kotlin/com/jzbrooks/vgo/core/transformation/RemoveRedundantPaintAttributesTests.kt
@@ -12,10 +12,10 @@ import com.jzbrooks.vgo.core.LinearGradient
 import com.jzbrooks.vgo.core.TileMode
 import com.jzbrooks.vgo.core.graphic.Circle
 import com.jzbrooks.vgo.core.graphic.Extra
+import com.jzbrooks.vgo.core.graphic.ForeignPaint
 import com.jzbrooks.vgo.core.graphic.Group
 import com.jzbrooks.vgo.core.graphic.PaintInheritance
 import com.jzbrooks.vgo.core.graphic.Path
-import com.jzbrooks.vgo.core.graphic.ForeignPaint
 import com.jzbrooks.vgo.core.util.element.createGraphic
 import com.jzbrooks.vgo.core.util.element.createPath
 import org.junit.jupiter.api.Test